created logger, test-helpful hardware emulator, added physical unit converters (to be check on real plant)

+++ RobotController.WpfClient/Charts/MeasurementModel.cs
@@ -5,6 +5,6 @@ namespace RobotController.WpfGui.Charts
     public class MeasurementModel
     {
         public DateTime DateTime { get; set; }
-        public short Value { get; set; }
+        public double Value { get; set; }
     }
 }
